lambda_btc_price_ingestor/requirements.txt
binance-connector
pandas
numpy
boto3
aws
boto
sagemaker
